requirements.txt
streamlit
numpy>=1.21.0
pillow
opencv-python>=4.5.0
keras
tensorflow>=2.12.0
imutils==0.5.4
matplotlib>=3.4.0
scipy>=1.6.0
scikit-learn>=0.24.0
